--- apps/email/README.md
@@ -1,1 +1,0 @@
-react-email doesn't exactly play nice with solid, so here we are.
