--- apps/email/emails/index.ts
@@ -1,2 +1,0 @@
-export { default as TenantAdminInviteEmail } from "./TenantAdminInvite";
-export { default as LoginCodeEmail } from "./LoginCode";
