--- apps/email/index.ts
@@ -1,1 +1,0 @@
-export type { RequestSchema } from "./app/route";
