--- apps/email/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "@mattrax/email",
-	"version": "0.1.0",
-	"types": "index.ts",
-	"private": true,
-	"scripts": {
-		"dev": "next dev --port 3005",
-		"build": "next build",
-		"start": "next start",
-		"email:build": "email build",
-		"email:dev": "email dev --port 3002",
-		"email:export": "email export"
-	},
-	"dependencies": {
-		"@react-email/button": "^0.0.14",
-		"@react-email/components": "0.0.16",
-		"@react-email/html": "^0.0.7",
-		"@react-email/render": "^0.0.12",
-		"@react-email/tailwind": "^0.0.15",
-		"@t3-oss/env-core": "^0.9.2",
-		"aws4fetch": "^1.0.18",
-		"next": "14.2.2",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-email": "2.1.1",
-		"zod": "^3.22.5"
-	},
-	"devDependencies": {
-		"@types/node": "^20.12.7",
-		"@types/react": "^18.2.79",
-		"@types/react-dom": "^18.2.25",
-		"typescript": "^5.4.5"
-	}
-}
