+++ apps/web/package.json
@@ -7,6 +7,9 @@
 		"build": "vinxi build",
 		"start": "vinxi start",
 		"stats": "pnpm build && open stats.html",
+		"email:build": "email build --dir ./src/emails",
+		"email:dev": "email dev --port 3002 --dir ./src/emails",
+		"email:export": "email export --dir ./src/emails",
 		"vbuild": "NITRO_PRESET=vercel_edge vinxi build",
 		"cbuild": "NITRO_PRESET=cloudflare_pages vinxi build"
 	},
@@ -16,14 +19,15 @@
 		"@kobalte/tailwindcss": "^0.9.0",
 		"@lucia-auth/adapter-drizzle": "^1.0.7",
 		"@mattrax/drizzle-to-rs": "workspace:*",
-		"@mattrax/email": "workspace:*",
 		"@mattrax/ms-graph": "workspace:*",
 		"@mattrax/trpc-server-function": "workspace:^",
 		"@mattrax/ui": "workspace:*",
 		"@microsoft/microsoft-graph-client": "^3.0.7",
 		"@paralleldrive/cuid2": "^2.2.2",
 		"@planetscale/database": "^1.16.0",
+		"@react-email/components": "0.0.16",
 		"@react-email/render": "^0.0.12",
+		"@react-email/tailwind": "^0.0.15",
 		"@solid-mediakit/trpc": "^3.0.2",
 		"@solid-primitives/context": "^0.2.3",
 		"@solid-primitives/date": "^2.0.21",
@@ -80,9 +84,12 @@
 		"@microsoft/microsoft-graph-types": "^2.40.0",
 		"@tailwindcss/forms": "^0.5.7",
 		"@types/prismjs": "^1.26.3",
+		"@types/react": "^18.2.79",
+		"@vitejs/plugin-react": "^4.2.1",
 		"autoprefixer": "^10.4.19",
 		"mysql2": "^3.9.7",
 		"postcss": "^8.4.38",
+		"react-email": "^2.1.1",
 		"rollup-plugin-visualizer": "^5.12.0",
 		"solid-devtools": "^0.30.1",
 		"tailwind-scrollbar": "^3.1.0",
