+++ apps/web/src/emails/TenantAdminInvite.email.tsx
@@ -1,3 +1,5 @@
+/** @jsxImportSource react */
+
 import {
 	Body,
 	Button,
