+++ apps/web/src/emails/UserEnrollmentInvite.email.tsx
@@ -1,3 +1,5 @@
+/** @jsxImportSource react */
+
 import {
 	Body,
 	Container,
