+++ apps/web/src/emails/index.ts
@@ -1,0 +1,3 @@
+export { default as TenantAdminInviteEmail } from "./TenantAdminInvite.email";
+export { default as LoginCodeEmail } from "./LoginCode.email";
+export { default as UserEnrollmentInviteEmail } from "./UserEnrollmentInvite.email";
